moved 2 projects in different folders
